requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.2
annotated-types==0.6.0
asgiref==3.8.1
asn1crypto==1.5.1
atpublic==4.1.0
attrs==23.1.0
av==10.0.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.2.1
crcmod==1.7
cryptography==38.0.4
cycler==0.11.0
dataclasses==0.6
Django==5.0
dnspython==2.6.1
dpcpp-cpp-rt==2024.1.0
easydict==1.10
exception==0.1.0
exceptiongroup==1.1.2
filelock==3.13.4
flit_core==3.6.0
fonttools==4.25.0
fsspec==2023.4.0
ftfy==6.2.0
future==0.18.3
h11==0.14.0
idna==3.4
imageio==2.34.1
importlib_metadata==7.1.0
intel-cmplr-lib-rt==2024.1.0
intel-cmplr-lic-rt==2024.1.0
intel-opencl-rt==2024.1.0
intel-openmp==2024.1.0
Jinja2==3.1.2
jmespath==0.10.0
kiwisolver==1.4.4
lazy_loader==0.4
llvmlite==0.42.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.0
mdurl==0.1.2
mkl==2024.1.0
mkl-fft==1.3.8
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==2.1.0
mmengine==0.10.3
mmsegmentation==1.2.2
model-index==0.1.11
mongoengine==0.28.2
mpmath==1.3.0
munkres==1.1.4
networkx==3.2.1
numba==0.59.1
numpy==1.23.2
opencv-python==4.7.0.72
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.38
ordered-set==4.1.0
oss2==2.17.0
outcome==1.2.0
packaging==22.0
pandas==2.2.2
Pillow==9.4.0
pip==24.0
platformdirs==4.2.0
ply==3.11
prettytable==3.10.0
protobuf==3.20.3
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.4
pywin32==306
PyYAML==6.0
regex==2024.4.16
requests==2.28.2
rich==13.4.2
scikit-image==0.23.2
scipy==1.13.0
selenium==4.10.0
setuptools==60.2.0
setuptools-scm==8.0.4
sip==6.6.2
six==1.16.0
sknw==0.15
sniffio==1.3.0
snowflake-connector-python==3.9.0
snowflake.core==0.7.0
snowflake._legacy==0.7.0
snowflake-snowpark-python==1.14.0
sortedcontainers==2.4.0
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tensorboardX==2.6
termcolor==2.4.0
tifffile==2024.4.24
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tornado==6.2
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.4.0
tzdata==2024.1
urllib3==1.26.14
wcwidth==0.2.13
wheel==0.38.4
wsproto==1.2.0
yacs==0.1.6
yapf==0.40.2
zipp==3.18.1
